requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2020.4.5.1
colorama==0.4.3
Django==3.0.3
django-crispy-forms==1.9.2
django-filter==2.3.0
easy-thumbnails==2.7
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
wincertstore==0.2
wrapt==1.11.2
